Move radio button to its own file, make some callbacks to handle the state changes in the parent
